Fix bug with incorrect `External function cannot have a body` error when calling external function
Function signature incorrectly assumed that type check is only called from the parent function and `TypeCheckState.curFunc` always points to the parent function. But it is possible for function signature type check to be initiated from the call expression, in which case the `TypeCheckState.curFunc` points to the caller function.
The fix is to move the check into the parent function.
Also did some refactoring.
Closes #16.