Fix #27. Correctly handle phi functions without users during register allocation.
Also, remove dead phi functions in DCE pass.

Here is what happens when bug is triggered:
First DCE removes some instructions. This leads to some phi functions not having any users. This makes liveness analysis emit empty range for those vregs (for example [144; 144)). Then register allocator allocates two intervals to the same register, and because overlap of [144; 144) and [144; 148) is not detected, they are both allocated to eax. Then move solver sees 2 writes to the same register and reports an error.